Decompose BaseUI
