+++ src/zrb/llm/ui/base/ui.py
@@ -42,12 +42,9 @@ from zrb.llm.tool_call import (
     default_response_handler,
 )
 from zrb.llm.ui.base.commands_mixin import CommandsMixin
+from zrb.llm.ui.base.replay_mixin import HistoryReplayMixin
+from zrb.llm.ui.base.system_info_mixin import SystemInfoMixin
 from zrb.llm.ui.multi_ui import MultiUI
-from zrb.llm.util.history_formatter import (
-    format_args,
-    format_timestamp,
-    truncate,
-)
 from zrb.session.any_session import AnySession
 from zrb.session.session import Session
 from zrb.task.any_task import AnyTask
@@ -64,7 +61,7 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class BaseUI(CommandsMixin):
+class BaseUI(CommandsMixin, HistoryReplayMixin, SystemInfoMixin):
     """Base class for LLM Chat UI implementations.
 
     This class provides the core chat functionality (message handling, command
@@ -677,107 +674,9 @@ class BaseUI(CommandsMixin):
                     # Event loop closed - exit
                     break
 
-    def _replay_history(self, messages: list) -> None:
-        """Render loaded conversation history through live-message paths.
-
-        Replays each ModelMessage so loaded sessions visually match a fresh
-        conversation: user lines and assistant headers render normally,
-        assistant text goes through markdown rendering, and tool calls/returns
-        use the same faint style as while streaming.
-        """
-        if not messages:
-            return
-        pending_tool_calls: dict[str, str] = {}
-        for msg in messages:
-            kind = getattr(msg, "kind", None)
-            timestamp = format_timestamp(getattr(msg, "timestamp", None))
-            ts_display = f"{timestamp} " if timestamp else ""
-            parts = getattr(msg, "parts", []) or []
-            if kind == "request":
-                self._replay_request_parts(parts, ts_display, pending_tool_calls)
-            elif kind == "response":
-                self.append_to_output(f"\n🤖 {ts_display}>>\n")
-                self._replay_response_parts(parts, pending_tool_calls)
-
-    def _replay_request_parts(
-        self,
-        parts: list,
-        ts_display: str,
-        pending_tool_calls: dict[str, str],
-    ) -> None:
-        """Render parts of a replayed ModelRequest (user/tool-return/retry)."""
-        for part in parts:
-            pkind = getattr(part, "part_kind", None)
-            if pkind == "tool-return":
-                self._replay_tool_return(part, pending_tool_calls)
-            elif pkind == "user-prompt":
-                content = str(getattr(part, "content", "") or "")
-                self.append_to_output(f"\n💬 {ts_display}>> {content.strip()}\n")
-            elif pkind == "retry-prompt":
-                content = str(getattr(part, "content", "") or "")
-                self.append_to_output(
-                    f"\n  🔄 Retry: {truncate(content, 200)}\n",
-                    kind="tool_call",
-                )
-            # system-prompt parts are skipped during replay
-
-    def _replay_response_parts(
-        self,
-        parts: list,
-        pending_tool_calls: dict[str, str],
-    ) -> None:
-        """Render parts of a replayed ModelResponse (thinking/text/tool-call)."""
-        width = self._get_output_field_width()
-        for part in parts:
-            pkind = getattr(part, "part_kind", None)
-            if pkind == "thinking":
-                content = str(getattr(part, "content", "") or "")
-                if content.strip():
-                    self.append_to_output(
-                        f"  💭 {truncate(content, 500)}", kind="thinking"
-                    )
-            elif pkind == "text":
-                content = str(getattr(part, "content", "") or "")
-                if content.strip():
-                    self.append_to_output("\n")
-                    self.append_to_output(
-                        render_markdown(
-                            content, width=width, theme=self._markdown_theme
-                        )
-                    )
-            elif pkind == "tool-call":
-                self._replay_tool_call(part, pending_tool_calls)
-
-    def _replay_tool_call(self, part, pending_tool_calls: dict[str, str]) -> None:
-        """Render a single replayed tool call."""
-        tool_name = getattr(part, "tool_name", None) or "unknown"
-        tool_call_id = getattr(part, "tool_call_id", None) or "?"
-        args_str = format_args(getattr(part, "args", None))
-        if tool_call_id and tool_name:
-            pending_tool_calls[tool_call_id] = tool_name
-        self.append_to_output(
-            f"  🧰 {tool_call_id} | {tool_name} {args_str}", kind="tool_call"
-        )
-
-    def _replay_tool_return(self, part, pending_tool_calls: dict[str, str]) -> None:
-        """Render a single replayed tool return."""
-        tool_call_id = getattr(part, "tool_call_id", None) or "?"
-        tool_name = (
-            getattr(part, "tool_name", None)
-            or pending_tool_calls.get(tool_call_id)
-            or "unknown"
-        )
-        outcome = getattr(part, "outcome", "success")
-        status_icon = "✅" if str(outcome) == "success" else "❌"
-        content = str(getattr(part, "content", "") or "")
-        self.append_to_output(
-            f"\n  🔠 {tool_call_id} | {tool_name} {status_icon}",
-            kind="tool_call",
-        )
-        body = truncate(content, 200)
-        if body.strip():
-            for line in body.split("\n")[:3]:
-                self.append_to_output(f"    {line}", kind="tool_call")
+    # History-replay rendering lives in HistoryReplayMixin (replay_mixin.py):
+    # _replay_history, _replay_request_parts, _replay_response_parts,
+    # _replay_tool_call, _replay_tool_return are inherited.
 
     def _submit_user_message(self, llm_task: AnyTask, user_message: str):
         # Check if we have a parent MultiUI to route through
@@ -918,69 +817,9 @@ class BaseUI(CommandsMixin):
             ui, call
         )  # --- SYSTEM INFO / TRIGGERS (Moved from UI) ---
 
-    async def _update_system_info(self):
-        """Update CWD and Git info."""
-        self._cwd = self._get_cwd_display()
-        branch, status = await self._get_git_info()
-        if branch:
-            self._git_info = f"{branch}{status}"
-        else:
-            self._git_info = "Not a git repo"
-        self.invalidate_ui()
-
-    def _get_cwd_display(self) -> str:
-        cwd = os.getcwd()
-        home = os.path.expanduser("~")
-        if cwd.startswith(home):
-            return "~" + cwd[len(home) :]
-        return cwd
-
-    async def _get_git_info(self) -> tuple[str, str]:
-        """Returns (branch_name, status_symbol)"""
-        try:
-            # Check branch
-            proc = await asyncio.create_subprocess_exec(
-                "git",
-                "rev-parse",
-                "--abbrev-ref",
-                "HEAD",
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-            )
-            stdout, _ = await proc.communicate()
-            if proc.returncode != 0:
-                return "", ""
-            branch = stdout.decode().strip()
-
-            # Check status (dirty or clean)
-            proc = await asyncio.create_subprocess_exec(
-                "git",
-                "status",
-                "--porcelain",
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-            )
-            stdout, _ = await proc.communicate()
-            is_dirty = bool(stdout.strip())
-
-            return branch, "*" if is_dirty else ""
-        except Exception:
-            return "", ""
-
-    async def _update_system_info_loop(self):
-        """Periodically update CWD and Git info."""
-        while True:
-            try:
-                await self._update_system_info()
-            except asyncio.CancelledError:
-                break
-            except Exception:
-                pass
-            try:
-                await asyncio.sleep(CFG.LLM_UI_LONG_STATUS_INTERVAL / 1000)
-            except RuntimeError:
-                # Event loop closed during shutdown
-                break
+    # System-info status (cwd/git) lives in SystemInfoMixin
+    # (system_info_mixin.py): _update_system_info, _get_cwd_display,
+    # _get_git_info, _update_system_info_loop are inherited.
 
     async def _trigger_loop(
         self,
